refactor(policy): split doctor policy domains
