+++ extensions/policy/src/doctor/metadata.test.ts
@@ -1,12 +1,8 @@
 // Policy doctor metadata tests cover rule metadata.
 import { describe, expect, it } from "vitest";
+import { CHECK_IDS, POLICY_CHECK_IDS } from "./check-ids.js";
 import { POLICY_FIX_METADATA_BY_CHECK_ID } from "./fix-metadata.js";
-import {
-  CHECK_IDS,
-  POLICY_CHECK_IDS,
-  POLICY_RULE_METADATA,
-  type PolicyRuleMetadata,
-} from "./metadata.js";
+import { POLICY_RULE_METADATA, type PolicyRuleMetadata } from "./metadata.js";
 
 const POLICY_FIX_METADATA = [...POLICY_FIX_METADATA_BY_CHECK_ID.values()];
 type PolicyFixMetadata = (typeof POLICY_FIX_METADATA)[number];
